requirements.txt
Flask
peewee
python-dotenv
mysql